config: Allow configuring the base runtime branch
You might want to build a dev runtime on top of a stable runtime, e.g.
org.gnome.Sdk//3.26. For that case, you need to be able to configure the
flatpak branch of the SDK, as well as the git branch of the git repo
where the manifest is located.

https://phabricator.endlessm.com/T18496