commands: Print warning about non-git sources
We currently only support git sources. If you want to hack on a module
that is specified as an archive source in the flatpak-builder manifest,
you can override it with a URL in your config file. This prints a
message explaining how to do that, since it wasn't documented anywhere.