config: Add option to add environment variables to module build
You can now add an extra section to your config file like this:

[mozjs52.extra_env]
AUTOCONF = autoconf-2.13

This will pass the given keys and values as environment variables in that
module's build options.

In order to make this work, we need to make config file keys
case-sensitive. (I didn't realize they were case-insensitive.) That is
technically a breaking change, so according to semver we'll bump the
version to 0.1.