ext: Make sure to get all config-opts
You can specify config-opts as a toplevel key in a module in the
manifest, or as a sub-key of build-options. When we modify the
config-opts from our config file, we need to make sure to combine both
of these config-opts and make sure the ones from our config file come
last.