ext: Add 'add_extensions' config key
This adds an 'add_extensions' config key which is used to specify
extensions that should be mounted during the build. For example, when
building packages that install gtk-doc files, it's necessary to have the
org.gnome.Sdk.Docs extension mounted. It can be specified as follows:

add_extensions =
    org.gnome.Sdk.Docs//3.26:share/runtime/docs

The triplet part is optional, the extension ID can be used by itself,
and in that case the branch will default to the runtime branch. The part
after the colon is the mount point.

https://phabricator.endlessm.com/T21545