Update example config file with new modules
There are new modules in the Endless SDK, so we should update the
example config file to allow hacking on them.

This takes all the non-dependency modules from the manifest and adds
them in the same order as they are ordered in the manifest, so
xapian-glib got reordered a bit.

https://phabricator.endlessm.com/T21272