config: strip ending '.git' string from git repo urls
There is an assumption that the last component of a git url matches
the directory name where it will be cloned. That is not true when
the URL ends with '.git' as it's typical on some hosting sites such
as gitlab.

This commit strips the ending '.git' from the url when getting the
checkout directory.